pyproject.toml
[tool.poetry]
name = "django-architecture"
version = "0.1.0"
description = ""
authors = ["billal <karimovbillal20002@gmail.com>"]
readme = "README.md"
packages = [{include = "django_architecture"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.5"
django-jazzmin = "^2.6.0"
django-cors-headers = "^4.0.0"
celery = {extras = ["redis"], version = "^5.2.7"}
django-filter = "^23.2"
pillow = "^9.5.0"
flower = "^1.2.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.1.0"}
twilio = "^8.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
